backend/requirements.txt
fastapi
uvicorn
playwright
httpx
python-dotenv
pydantic
beautifulsoup4
python-multipart
firebase-admin
